feat(exercises): allow users to create new exercises from the exercises tab
