+++ src/app/(tabs)/exercises/_layout.tsx
@@ -14,6 +14,7 @@ export default function ExercisesLayout() {
       }}
     >
       <Stack.Screen name="index" />
+      <Stack.Screen name="new" />
       <Stack.Screen name="[id]" />
     </Stack>
   );
